requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

sqlalchemy==2.0.34
psycopg2-binary==2.9.9

pydantic==2.9.2
python-jose==3.3.0

argon2-cffi==23.1.0

python-dotenv==1.0.1

# Utilidades
passlib==1.7.4
